+++ chaoslib/discovery/__init__.py
@@ -1,3 +1,6 @@
 # -*- coding: utf-8 -*-
 from chaoslib.discovery.discover import discover, \
     discover_actions, discover_probes, initialize_discovery_result
+
+__all__ = ["discover", "discover_actions", "discover_probes",
+           "initialize_discovery_result"]