+++ chaoslib/provider/process.py
@@ -2,7 +2,6 @@
 import itertools
 import os
 import os.path
-import shlex
 import shutil
 import subprocess
 from typing import Any