+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 [Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.0.0rc3...HEAD
 
+- Delint to clean things up before 1.0
+
 ## [1.0.0rc3][] - 2019-01-29
 
 [1.0.0rc3]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.0.0rc2...1.0.0rc3